Moved Header code from main frame to seperate component
